Create pH_control.ino
The pH electrode probe is connected to a  GAOHOU Value Detect Sensor Module (reads pH values 0-14), and the Creality stepper motor is connected to an L298N motor drive controller board module. The sensor module and the L298N motor drive controller board module are connected to an ELEGOO UNO R3 microcontroller board. When the electrode probe is submerged in a liquid, the sensor module reads the pH, and if the pH reading is less than 7, the L298N motor drive controller board module prompts the stepper motor to rotate 0.5 revolutions clockwise. If the pH reading is more than 7, the motor rotates 0.5 revolutions counterclockwise. The rotation of the motor will seal/uncover the pH-neutralizing agent which will bring the pH level back to normal.