Create pH_control_solenoid.ino
A pH electrode probe is connected to a GAOHOU value detect sensor module (reads pH values 0-14). Three relays: relay1, relay2, and relay3 are connected to solenoid1, solenoid2, and solenoid3 respectively. The Value Detect Sensor module and the three relays are connected to an ELEGOO UNO R3 microcontroller board. When the electrode probe is submerged in a liquid, the GAOHOU value detect sensor module reads the pH. It displays the pH on the serial monitor for debugging purposes. Whenever the pH reading is less than 7, relay1 turns on and solenoid1 (basic agent) is turned on for 2 seconds. If the pH reading is more than 7, relay2 turns on and solenoid2 (acidic agent) is turned on for 2 seconds. Relay3 causes solenoid3 (nutrients) to turn on for 5 seconds once every 2 hours.